Allow for dynamic return sizes from procedures.
Previously, the expeted return size had to be included with any system
calls as the return size needed to be known by the kernel. This
commit changes the kernel (and the example procedures) to use
RETURNDATASIZE and RETURNDATACOPY. Using these opcodes, the size
of the return data is determined after the call and can be of any size.