Alpha Implementation of Specification (#129)
* Remove last references to push and delete caps

* Update proc table and cap formats (no minting)

This updates the procedure list and procedure heap to use the correct
formats. It does not attempt to address minting and delegation and
simply allows anything to be registered. Minting and delegation will be
addressed in a later PR.

* Test updates for new formats.

* Implement maximal caps

* Implement subsets

* Test registration of Call caps

* Test registration of Write caps

* Test registration of Log caps

* Various test fixes and redesign

* Critical fix for Delete cap

* Minor test fixes

* Remove Procedure struct from registration path

* Add some utility functions for tests

May seem a little verbose but it's a first step in reducing the
boilerplate of tests. Currently we've avoided that, but the test have
grown to large and we need to start abstracting bits.

* Test procedure re-registration

This required some error code plumbing to be done. I also removed one
of the checks that returned an error code if the procedure name already
exists. This way we only check that problem once rather than having
unreachable code.

* Remove erroneous .only

* Correct error code

* Test zeroing of clists at registration (see #129 for: at deletion).

This is a quick patch that sets the clists to zero length before
registering. The final choice for how we deal with clist deletion is to
be made in #129.

* Remove old hack for tests

There use to be a special cap class to monkey patch the fact that the
capIndex value it not passed through correctly. This is now fixed.

* Fix SetEntry tests

* Missing changes from last commit

* Restructure files and inheritance

* Update tests to restructure

* Added missing files

* Remove deprecated error values

* Set the kernel address in preparation for execution guard

* Constructor bootstrap proof-of-concept.

This commit gives TestKernel a constructor which takes an initial entry
procedure. This required test changes throughout. The implementation
will need to be updated to have a better architecture, but this
implements the concepts and updates tests for compatibility.

The next steps will be to move it to a level lower then TestKernel and
abstract some of the initial cap setup. The concept is documented in
in issue #130.

* Add KernelInstance.sol

This moves the constructor and bootstrap mechanism to another layer in
KernelInstance.sol, as per #130.

* Set the kernel address in the constructor.

* Address comments

* Add missing arrow to architecture diagram