Fix link errors.
When we are compiling to WASM, unresolved references are left as
(import) expressions. However, under any other target symbols will have
to be linked for EVM functions (blocknumber, create, etc.). Therefore,
when we are not compiling for WASM (be it test, realse, whatever) we
want to link in dummy functions. pwasm_test provides all the builtins
provided by parity, while cap9_test covers the few that we have
implemented ourselves.